Documents/DV-2027-ETH/cpanel_deployment/requirements.txt
django-crispy-forms
crispy-bootstrap5
django-countries
django-phonenumber-field
phonenumbers
python-telegram-bot
Pillow
whitenoise
gunicorn
python-dotenv
python-decouple
